wip: Update ProjectStorage display in project page
