+++ client/src/features/ProjectPageV2/ProjectPageContent/ProjectStorage/projectStorage.constants.ts
@@ -1,0 +1,5 @@
+export const PROJECT_STORAGE_MIN_GB = 1;
+export const PROJECT_STORAGE_MAX_GB = 10;
+export const PROJECT_STORAGE_STEP_GB = 1;
+export const PROJECT_STORAGE_DEFAULT_GB = 1;
+export const PROJECT_STORAGE_DEFAULT_MOUNT_PATH = "store";
